moved ThemeProvider to App.tsx
