+++ src/App.tsx
@@ -1,15 +1,19 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Car  from "./pages/Mykyta/Car";
+import Car  from "./pages/Car";
 import Layout from "./pages/Layout";
 import Rent from "./pages/Rent";
 import Home from "./pages/Home";
 import Error from "./pages/Error";
+import { CssBaseline, ThemeProvider } from "@mui/material";
+import theme from "./theme";
 
 import './App.css';
 
 function App() {
 
   return (
+    <ThemeProvider theme={theme}>
+    <CssBaseline/>
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Layout />} >
@@ -20,6 +24,7 @@ function App() {
         </Route>
       </Routes>
     </BrowserRouter>
+    </ThemeProvider>
   );
 }
