+++ src/pages/Error.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
-import { CssBaseline, ThemeProvider} from '@mui/material'
 import Typography from '@mui/material/Typography'
 import Container from '@mui/material/Container'
-import theme from '../theme';
+
 
 
 const Error: React.FC = () => {
     return (
-        <ThemeProvider theme={theme}>
-            <CssBaseline />
             <Container sx={{width:"100%" , display:"flex" , flexDirection:"column", alignItems: "center"}}>
                 <Typography variant="h1" component="h2" gutterBottom>
                     Error 404
@@ -17,7 +14,6 @@ const Error: React.FC = () => {
                     {'This path is incorrect. Please, go back to the main page.'}
                 </Typography>
             </Container>
-        </ThemeProvider>
     );
 }
