+++ src/pages/Layout.tsx
@@ -21,7 +21,7 @@ function Layout() {
             <AppBar position="static" sx={NavBar} >
                 <Toolbar>
                     <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-                        Rent Car
+                        Delta
                     </Typography>
                     <Button sx={Buttons} color="inherit" component={Link} to="/">Main page</Button>
                     <Button sx={Buttons} color="inherit" component={Link} to="/rent-car-form">Rent Car</Button>
